@martinholters 2024-11-19
I've never looked at the algorithm/its implementation in any detail and don't have the bandwidth to do so in the near future. Wanna open an issue about it so we don't forget?